Low level programming

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -7,19 +7,19 @@
 int main(void)
 {
 	int i;
-	long int f[98];
+	long double f[98];
 
 	f[0] = 1;
 	f[1] = 2;
-	printf("%ld, %ld, ", f[0], f[1]);
+	printf("%Lf, %Lf, ", f[0], f[1]);
 
 	for (i = 2; i < 98; i++)
 	{
 		f[i] = f[i - 1] + f[i - 2];
 		if (i == 97)
-			printf("%ld\n", f[i]);
+			printf("%Lf\n", f[i]);
 		else
-			printf("%ld, ", f[i]);
+			printf("%Lf, ", f[i]);
 	}
 
 	return (0);
